requirements.txt
amqp==5.3.1
appdirs==1.4.4
asgiref==3.11.0
billiard==4.2.4
celery==5.6.1
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==6.0
django-autoslug==1.9.9
django-cors-headers==4.9.0
django-extensions==4.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
fs==2.4.16
inflection==0.5.1
kombu==5.6.2
packaging==25.0
prompt_toolkit==3.0.52
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.1.0
setuptools==79.0.1
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
unicode-slugify==0.1.5
Unidecode==1.4.0
uuid==1.30
vine==5.1.0
wcwidth==0.2.14
